api/requirements.txt
openai
python-dotenv
pymysql
